Code cleanup
Also changed the continue game yes/no buttons to confirmation buttons
built into the dialog. Also added a background to the sign out button.

Moved many more strings to xml. This is all of the hardcoded strings
that the inspector found but there are probably a few more.

See #58 Move strings to xml

+++ app_2048/src/main/java/com/tytanapps/game2048/GameData.java
@@ -103,7 +103,7 @@ public class GameData implements java.io.Serializable {
 
             byte[] imageByteArray = new byte[imageByteList.size()];
             for(int i = 0; i < imageByteList.size(); i++)
-                imageByteArray[i] = imageByteList.get(i).byteValue();
+                imageByteArray[i] = imageByteList.get(i);
 
             BitmapFactory.Options opt = new BitmapFactory.Options();
             opt.inPreferredConfig = Bitmap.Config.ARGB_8888;
